+++ test/src/tests/widgets/test_label.py
@@ -1,35 +1,24 @@
-from pytest import approx, fixture
+from pytest import approx, fixture, mark
 
 import toga
 
-from ..test_data import COLORS, TEXTS
-from ..test_utils import set_get
+from .common import test_background_color, test_color, test_text  # noqa: F401
 
 
 @fixture
-async def new_widget():
+async def widget():
     return toga.Label("")
 
 
-async def test_text(widget, probe):
-    for text in TEXTS:
-        set_get(widget, "text", text)
-        assert probe.text == text
-
-
-async def test_color(widget, probe):
-    for color in COLORS:
-        widget.style.color = color
-        for component in ["r", "g", "b"]:
-            assert getattr(probe.color, component) == getattr(color, component)
-        assert probe.color.a == approx(color.a, abs=(1 / 256))
-
-
+# TODO: a `width` test, for any widget whose width depends on its text.
+@mark.skip("Fails on Windows, probably because of #1289")
 async def test_multiline(widget, probe):
     def make_lines(n):
         return "\n".join(f"line{i}" for i in range(n))
 
     widget.text = make_lines(1)
+    # TODO: Android at least will need an `await` after each text change, to give the
+    # native layout a chance to update.
     line_height = probe.height
 
     widget.text = make_lines(2)
